requirements.txt
Django==5.2.10
djangorestframework==3.15.2
drf-spectacular==0.27.2
django-cors-headers==4.4.0
mysqlclient==2.2.4
python-dotenv==1.0.1
Pillow==10.4.0
